+++ src/components/navBar/NavBar.js
@@ -21,13 +21,6 @@ class NavBar extends Component {
     });
   };
 
-
-  handleLogout() {
-    axios.get(`${authURL}/logout`).then((response) => {
-      this.props.checkLogout(); // Handle loggin out of a user, dumping token
-    });
-  }
-
   headerButton() {
     if (this.props.isLoggedIn === false) {
       return <button type="button" onClick={this.toggle} className="btn edit-button ml-auto">Sign Up</button>
@@ -35,17 +28,16 @@ class NavBar extends Component {
       return (
         <div>
           <a href="/"><button type="Link" onClick={this.taskView} className="btn edit-button ml-auto">Active Tasks </button></a>
-          <button type="button" onClick={(e) => this.handleLogout()} className="btn edit-button ml-auto">Log out</button>
+          <a href="/"><button type="button" onClick={(e) => this.props.checkLogout()} className="btn edit-button ml-auto">Log out</button></a>
         </div>
       )
     } else if (this.props.isLoggedIn === true && this.props.location.pathname === '/')
       return (
         <div>
           <a href="/archived"><button type="Link" onClick={this.taskView} className="btn edit-button ml-auto">Archived Tasks </button></a>
-          <button type="button" onClick={(e) => this.handleLogout()} className="btn edit-button ml-auto">Log out</button>
+          <button type="button" onClick={(e) => this.props.checkLogout()} className="btn edit-button ml-auto">Log out</button>
         </div>
       )
-
   }
 
   render() {
@@ -58,14 +50,6 @@ class NavBar extends Component {
           </button>
           <div className="collapse navbar-collapse" id="navbarNav">
             { this.headerButton() }
-          </div>
-          <div>
-            {
-              this.props.match.path === '/tasks' &&
-              <a href="/archived"><button type="Link" onClick={this.taskView} className="btn edit-button ml-auto">
-                Archived Task
-              </button></a>
-            }
           </div>
         </nav>
         <SignUpModal isOpen={this.state.modal} onClick={this.toggle} />
